# maryema_api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Maryema API"
  version: "1.0"
  description: "API documentation for Maryema application"
servers:
  - url: "https://127.0.0.1:8000/api"
    description: "API base URL"
...
components:
  schemas:
    Profile:
      type: object
      properties:
        role:
          type: string
          enum: [admin, customer, provider]
          default: customer
        phone_number:
          type: string
          pattern: ^01[012]\d{8}$
        avatar:
          type: string
          format: uri
          writeOnly: true
          description: the avatar image field for create and update operations
        avatar_url:
          type: string
          format: uri
          pattern: ^https://avatar\.iran\.liara\.run/public/[5-9][0-9]$
          readOnly: true
          description: Avatar URL endpoint for src attribute in img tag
        note:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        url:
          type: string
          format: uri
          pattern: ^http://127\.0\.0\.1:8000/api/admin/users/[0-9]$
          readOnly: true
          description: User profile details URL endpoint
        date_joined:
          type: string
          format: date-time
          readOnly: true
        last_login:
          type: string
          format: date-time
          readOnly: true
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
        first_name:
          type: string
        last_name:
          type: string
        is_active:
          type: boolean
        profile:
          $ref: "#/components/schemas/Profile"
      required:
        - username
        - email
        - password
...
paths:
  /me/:
    get:
      tags: [Current User]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required

    patch:
      tags: [Current User]
      summary: Update current user profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid data
        "401":
          description: Authentication required

    delete:
      tags: [Current User]
      summary: Delete current user account
      responses:
        "204":
          description: Account deleted
        "401":
          description: Authentication required
...